src/main/java/es/plexus/hopes/hopesback/controller/RoleController.java
package es.plexus.hopes.hopesback.controller;

import es.plexus.hopes.hopesback.controller.model.RoleDTO;
import es.plexus.hopes.hopesback.service.RoleService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping(RoleController.ROLE_MAPPING)
public class RoleController {

	static final String ROLE_MAPPING = "/role";

	private final RoleService roleService;

	@Autowired
	public RoleController(RoleService roleService) {
		this.roleService = roleService;
	}

	@GetMapping
	public List<RoleDTO> getAllRoles() {
		return roleService.getAllRoles();
	}
}
